+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - [cronos#1163](https://github.com/crypto-org-chain/cronos/pull/1163) Support stateful precompiled contract for ica.
 - [cronos#837](https://github.com/crypto-org-chain/cronos/pull/837) Support stateful precompiled contract for bank.
 - [cronos#1184](https://github.com/crypto-org-chain/cronos/pull/1184) Update ibc-go to `v7.3.1`.
+- [cronos#1185](https://github.com/crypto-org-chain/cronos/pull/1185) Support ibc callback.
 
 ### Bug Fixes